Minor changes to Virtualization Framework
