+++ MacMulator/Utils/VirtualizationFrameworkMacOSSupport.swift
@@ -9,7 +9,7 @@ import Foundation
 import Virtualization
 
 @available(macOS 12.0, *)
-class VirtualizationFrameworkUtils {
+class VirtualizationFrameworkMacOSSupport {
     
     static let RESTORE_IMAGE_NAME = "macos-restore-image.ipsw"
     static let AUXILIARY_STORAGE_NAME = "auxiliary-storage"
@@ -61,7 +61,7 @@ class VirtualizationFrameworkUtils {
         
         virtualMachineConfiguration.storageDevices = [MacOSVirtualMachineConfigurationHelper.createBlockDeviceConfiguration(path: Utils.findMainDrive(vm.drives)!.path)]
         virtualMachineConfiguration.networkDevices = [MacOSVirtualMachineConfigurationHelper.createNetworkDeviceConfiguration()]
-        virtualMachineConfiguration.pointingDevices = [MacOSVirtualMachineConfigurationHelper.createPointingDeviceConfiguration()]
+        virtualMachineConfiguration.pointingDevices = MacOSVirtualMachineConfigurationHelper.createPointingDeviceConfigurations()
         virtualMachineConfiguration.keyboards = [MacOSVirtualMachineConfigurationHelper.createKeyboardConfiguration()]
         virtualMachineConfiguration.audioDevices = [MacOSVirtualMachineConfigurationHelper.createAudioDeviceConfiguration()]
         
@@ -72,9 +72,9 @@ class VirtualizationFrameworkUtils {
     
     fileprivate static func createMacPlatformConfiguration(vm: VirtualMachine, macOSConfiguration: VZMacOSConfigurationRequirements?) -> VZMacPlatformConfiguration {
         let macPlatformConfiguration = VZMacPlatformConfiguration()
-        let auxiliaryStorageURL = URL(fileURLWithPath: vm.path + "/" + VirtualizationFrameworkUtils.AUXILIARY_STORAGE_NAME + "-0")
-        let machineIdentifierURL = URL(fileURLWithPath: vm.path + "/" + VirtualizationFrameworkUtils.MACHINE_IDENTIFIER_NAME + "-0")
-        let hardwareModelURL = URL(fileURLWithPath: vm.path + "/" + VirtualizationFrameworkUtils.HARDWARE_MODEL_NAME + "-0")
+        let auxiliaryStorageURL = URL(fileURLWithPath: vm.path + "/" + VirtualizationFrameworkMacOSSupport.AUXILIARY_STORAGE_NAME + "-0")
+        let machineIdentifierURL = URL(fileURLWithPath: vm.path + "/" + VirtualizationFrameworkMacOSSupport.MACHINE_IDENTIFIER_NAME + "-0")
+        let hardwareModelURL = URL(fileURLWithPath: vm.path + "/" + VirtualizationFrameworkMacOSSupport.HARDWARE_MODEL_NAME + "-0")
         
         if macOSConfiguration != nil {
             // Create resources using provided configuration
@@ -119,94 +119,4 @@ class VirtualizationFrameworkUtils {
     
 #endif
     
-    @available(macOS 13.0, *)
-    static func createLinuxVirtualMachineData(vm: VirtualMachine) {
-        
-        _ = createLinuxPlatformConfiguration(vm: vm, isInstalling: true)
-        _ = createLinuxBootloader(vm: vm, isInstalling: true)
-    }
-    
-    @available(macOS 13.0, *)
-    static func decodeLinuxVirtualMachine(vm: VirtualMachine, installMedia: String) -> VZVirtualMachine {
-        let configuration = setupLinuxVirtualMachine(vm: vm, installMedia: installMedia)
-        
-        let virtualMachine = VZVirtualMachine(configuration: configuration, queue: .main)
-        return virtualMachine;
-    }
-    
-    @available(macOS 13.0, *)
-    fileprivate static func setupLinuxVirtualMachine(vm: VirtualMachine, installMedia: String) -> VZVirtualMachineConfiguration {
-        let virtualMachineConfiguration = VZVirtualMachineConfiguration()
-        virtualMachineConfiguration.platform = createLinuxPlatformConfiguration(vm: vm, isInstalling: false)
-        virtualMachineConfiguration.cpuCount = vm.cpus
-        virtualMachineConfiguration.memorySize = UInt64(vm.memory) * (1024 * 1024)
-        virtualMachineConfiguration.bootLoader = createLinuxBootloader(vm: vm, isInstalling: false)
-        
-        let resolution = Utils.getResolutionElements(vm.displayResolution)
-        
-        virtualMachineConfiguration.graphicsDevices = [LinuxVirtualMachineConfigurationHelper.createGraphicsDeviceConfiguration(
-            witdh: resolution[0],
-            height: resolution[1])]
-        
-        let disksArray = NSMutableArray()
-        if installMedia != "" {
-            disksArray.add(LinuxVirtualMachineConfigurationHelper.createUSBMassStorageDeviceConfiguration(installMedia))
-        }
-        disksArray.add(LinuxVirtualMachineConfigurationHelper.createBlockDeviceConfiguration(path: Utils.findMainDrive(vm.drives)!.path))
-        guard let disks = disksArray as? [VZStorageDeviceConfiguration] else {
-            fatalError("Invalid disksArray.")
-        }
-        
-        virtualMachineConfiguration.storageDevices = disks
-        virtualMachineConfiguration.networkDevices = [LinuxVirtualMachineConfigurationHelper.createNetworkDeviceConfiguration()]
-        virtualMachineConfiguration.pointingDevices = [LinuxVirtualMachineConfigurationHelper.createPointingDeviceConfiguration()]
-        virtualMachineConfiguration.keyboards = [LinuxVirtualMachineConfigurationHelper.createKeyboardConfiguration()]
-        virtualMachineConfiguration.audioDevices = [LinuxVirtualMachineConfigurationHelper.createAudioDeviceConfiguration()]
-        virtualMachineConfiguration.consoleDevices = [LinuxVirtualMachineConfigurationHelper.createSpiceAgentConsoleDeviceConfiguration()]
-        
-        try! virtualMachineConfiguration.validate()
-        
-        return virtualMachineConfiguration
-    }
-    
-    @available(macOS 13.0, *)
-    fileprivate static func createLinuxPlatformConfiguration(vm: VirtualMachine, isInstalling: Bool)  -> VZGenericPlatformConfiguration {
-        let linuxPlatformConfiguration = VZGenericPlatformConfiguration()
-        
-        let machineIdentifierURL = URL(fileURLWithPath: vm.path + "/" + VirtualizationFrameworkUtils.MACHINE_IDENTIFIER_NAME + "-0")
-        
-        if isInstalling {
-            linuxPlatformConfiguration.machineIdentifier = VZGenericMachineIdentifier()
-            try! linuxPlatformConfiguration.machineIdentifier.dataRepresentation.write(to: machineIdentifierURL)
-        } else {
-            guard let machineIdentifierData = try? Data(contentsOf: machineIdentifierURL) else {
-                fatalError("Failed to retrieve machine identifier data.")
-            }
-            guard let machineIdentifier = VZGenericMachineIdentifier(dataRepresentation: machineIdentifierData) else {
-                fatalError("Failed to create machine identifier.")
-            }
-            linuxPlatformConfiguration.machineIdentifier = machineIdentifier
-        }
-        
-        return linuxPlatformConfiguration
-    }
-    
-    @available(macOS 13.0, *)
-    fileprivate static func createLinuxBootloader(vm: VirtualMachine, isInstalling: Bool)  -> VZEFIBootLoader {
-        let bootloader = VZEFIBootLoader()
-        
-        let efiVariableStorePath = vm.path + "/" + VirtualizationFrameworkUtils.EFI_VARIABLE_STORE_NAME + "-0"
-        
-        if isInstalling {
-            bootloader.variableStore = LinuxVirtualMachineConfigurationHelper.createEFIVariableStore(path: efiVariableStorePath)
-        } else {
-            if !FileManager.default.fileExists(atPath: efiVariableStorePath) {
-                fatalError("EFI variable store does not exist.")
-            }
-            
-            bootloader.variableStore = VZEFIVariableStore(url: URL(fileURLWithPath: efiVariableStorePath))
-        }
-        
-        return bootloader
-    }
 }
